+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to `laravel-zammad` will be documented in this file.
 
 ### Unreleased
 
+- ⚠️ [Breaking Change] Changed `user_id` to `customer_id` in the ticket DTO.
 - Added sender_id for the comment DTO.
 
 ## 0.0.1 - 2021-04-21